python-quantum/requirements.txt
flask==2.3.3
qiskit==0.44.1
qiskit-aer==0.12.2
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
